+++ electron.vite.config.ts
@@ -47,6 +47,7 @@ export default defineConfig({
             if (
               id.includes('node_modules/react-markdown') ||
               id.includes('node_modules/remark') ||
+              id.includes('node_modules/rehype') ||
               id.includes('node_modules/mdast') ||
               id.includes('node_modules/micromark') ||
               id.includes('node_modules/unist') ||
@@ -72,6 +73,12 @@ export default defineConfig({
               id.includes('node_modules/is-plain-obj')
             ) {
               return 'markdown'
+            }
+            if (id.includes('node_modules/katex')) {
+              return 'katex'
+            }
+            if (id.includes('node_modules/mermaid')) {
+              return 'mermaid'
             }
             return undefined
           }